Final update for movie migration and button fix
Changed the AddMovieDataSet NumberInStock field from byte to int to
match model & moved add new buttons out of movie and customer index
pages

+++ Vidly/Migrations/201704062054109_AddMoviesDataSet.cs
@@ -16,7 +16,7 @@ namespace Vidly.Migrations
                     GenreId = c.Byte(nullable: false),
                     DateAdded = c.DateTime(nullable: false),
                     ReleaseDate = c.DateTime(nullable: false),
-                    NumberInStock = c.Byte(nullable: false),
+                    NumberInStock = c.Int(nullable: false),
                 })
                 .PrimaryKey(t => t.Id)
                 .ForeignKey("dbo.Genres", t => t.GenreId, cascadeDelete: true)
